+++ tensorflow_toolkit/text_detection/.gitignore
@@ -1,0 +1,6 @@
+__pycache__/
+.idea/
+init_weights
+data
+model
+venv_mo
