+++ tensorflow_toolkit/text_detection/README.md
@@ -1,8 +1,12 @@
-# Text Detection in TensorFlow 2.0
+# Text Detection in TensorFlow
 
 This repository contains inference and training code for [PixelLink](https://arxiv.org/abs/1801.01315)-like model
 networks. Models code is designed to enable export to frozen graph
 and inference on CPU via OpenVINO. You can refer to [Original implementation](https://github.com/ZJULearning/pixel_link) as well.
+
+Trained models: [link](https://download.01.org/opencv/openvino_training_extensions/models/text_detection/text_detection.tar.gz)
+
+![](text-detection.jpg)
 
 ## Setup
 
@@ -15,20 +19,19 @@ and inference on CPU via OpenVINO. You can refer to [Original implementation](ht
 
 ### Installation
 
-It is recommended to use virtual environment, to do that please install `virtualenv`
+Create virtual environment
 ```bash
-sudo apt update
-sudo apt install python3-dev python3-pip graphviz
-sudo pip3 install -U virtualenv  # system-wide install
-
-virtualenv -p python3 ./venv
-source ./venv/bin/activate
+virtualenv venv -p python3 --prompt="(td)"
 ```
 
-To install required dependencies run
-
+2. Activate virtual environment and setup OpenVINO variables
 ```bash
-pip install -r requirements.txt
+. venv/bin/activate
+```
+
+3. Install the module
+```bash
+pip3 install -e .
 ```
 
 ## <a name="Dataset"> Dataset </a>
@@ -54,13 +57,18 @@ You can use `annotation.py` module to convert listed above datasets to internal 
 To convert toy dataset located in `./data` folder please run:
 
 ```bash
-python annotation.py --type toy --images data --out_annotation data/annotation.json --imshow_delay 3
+python tools/annotation.py \
+  --type toy \
+  --images data \
+  --out_annotation data/annotation.json
 ```
 
 To create TFRecordDataset please run:
 
 ```bash
-python create_dataset.py --input_datasets data/annotation.json --output data/annotation.tfrecord --imshow_delay 1
+python tools/create_dataset.py \
+  --input_datasets data/annotation.json \
+  --output data/annotation.tfrecord
 ```
 
 ## Training
@@ -74,41 +82,49 @@ To run training you need to specify :
 * `epochs_per_evaluation` - how often model will be saved/evaluated.
 
 ```bash
-$ python train.py \
---learning_rate 0.001 \
---train_dir model \
---train_dataset ./data/annotation.tfrecord \
---epochs_per_evaluation 100 \
---test_dataset ./data/annotation.tfrecord \
---model_type mobilenet_v2_ext \
---config configs/config.yaml
+python tools/train.py \
+  --learning_rate 0.001 \
+  --train_dir model \
+  --train_dataset data/annotation.tfrecord \
+  --epochs_per_evaluation 100 \
+  --test_dataset data/annotation.tfrecord \
+  --model_type mobilenet_v2_ext \
+  --config configs/config.yaml
 ```
 
 You can add one more parameter such as:
 * `weights` - weights of pretrained model. That can give your faster convergence and better model.
 
 ```bash
-$ python train.py \
---learning_rate 0.001 \
---train_dir model \
---train_dataset ./data/annotation.tfrecord \
---epochs_per_evaluation 100 \
---test_dataset ./data/annotation.tfrecord \
---model_type mobilenet_v2_ext \
---config configs/config.yaml
---weights some_pretrained_model.saved_weights
+python tools/train.py \
+  --learning_rate 0.0001 \
+  --train_dir model \
+  --train_dataset data/annotation.tfrecord \
+  --epochs_per_evaluation 100 \
+  --test_dataset data/annotation.tfrecord \
+  --model_type mobilenet_v2_ext \
+  --config configs/config.yaml \
+  --weights init_weights/model_mobilenet_v2_ext/weights/model-523.save_weights
 ```
 
+## Evaluation
 
-## Export TF-2.0 models to OpenVINO (IR)
+```bash
+python tools/test.py \
+  --config model/configuration.yaml \
+  --dataset data/annotation.tfrecord \
+  --weights model/weights/model-500.save_weights
+```
+
+## Export models to OpenVINO (IR)
 
 * First step is to freeze your model. You will need configuration file that appears in `train_dir` during training.
 
 ```bash
-python freeze.py \
---resolution 1280 768 \
---config configuration.yaml \
---weights some_pretrained_model.saved_weights
+python tools/export.py \
+  --resolution 1280 768 \
+  --config model/configuration.yaml \
+  --weights model/weights/model-500.save_weights
 ```
 
 It will print information about frozen model and how to get IR:
@@ -119,25 +135,30 @@ Operations number: 51.075934092 GFlops
 Output tensor names for using in InferenceEngine:
      model/link_logits_/add
      model/segm_logits/add
-Run model_optimizer to get IR: mo.py --input_model /tmp/tmp4vhg5k2i/frozen_graph.pb --reverse_input_channels
+Run model_optimizer to get IR: mo.py --input_model model/weights/export/frozen_graph.pb --reverse_input_channels
 ```
 
 * Second step is to run `mo.py`
 
-You might need to install TF1.13 to run following line.
+**NOTE** You need to install TF1.13 to use model optimizer.
 
+1. Create and activate new virtual environment
 ```bash
-mo.py --input_model /tmp/tmp4vhg5k2i/frozen_graph.pb --reverse_input_channels
+virtualenv venv_mo -p python3 --prompt="(td-mo)"
+. venv_mo/bin/activate
 ```
 
-## Evaluation in TF-2.0
-
+3. Install modules and activate environment for OpenVINO
 ```bash
-python test.py \
---config configuration.yaml \
---dataset dataset.tfrecord \
---weights some_pretrained_model.saved_weights
+pip3 install -r requirements-mo.txt
+source /opt/intel/openvino/bin/setupvars.sh
 ```
+
+4. Run model optimizer tool to export frozen graph to IR
+```bash
+mo.py --input_model model/weights/export/frozen_graph.pb --reverse_input_channels
+```
+
 
 ## Demo in OpenVINO
