+++ tensorflow_toolkit/text_detection/configs/config.yaml
