+++ tensorflow_toolkit/text_detection/data/gt_img_1.txt
