+++ tensorflow_toolkit/text_detection/data/gt_img_2.txt
