+++ tensorflow_toolkit/text_detection/data/gt_img_4.txt
