+++ tensorflow_toolkit/text_detection/data/img_1.jpg
