+++ tensorflow_toolkit/text_detection/data/img_2.jpg
