+++ tensorflow_toolkit/text_detection/data/img_3.jpg
