+++ tensorflow_toolkit/text_detection/data/img_4.jpg
