+++ tensorflow_toolkit/text_detection/requirements-mo.txt
@@ -1,0 +1,5 @@
+tensorflow==1.13.1
+networkx>=1.11
+numpy>=1.12.0
+test-generator==0.1.1
+defusedxml==0.5.0
