+++ tensorflow_toolkit/text_detection/requirements.txt
@@ -8,6 +8,7 @@ Keras-Applications==1.0.7
 Keras-Preprocessing==1.0.9
 Markdown==3.1
 numpy==1.16.2
+opencv-python==4.1.0.25
 Polygon3==3.0.8
 protobuf==3.7.1
 PyYAML==5.1
