+++ tensorflow_toolkit/text_detection/text_detection/dataset.py
