+++ tensorflow_toolkit/text_detection/text_detection/model.py
