--- tensorflow_toolkit/text_detector/.gitignore
@@ -1,2 +1,0 @@
-__pycache__/
-.idea/
